Simplify conditional statements
In python, comparing to booleans is usually frowned upon.
To test if a variable is of NoneType, one should use `is None` or `is not None`, not `==` or `!=`.
Conditions do not need to be parenthesized.
And iterating over a list to find something is the same as using the contains magic method by asking `if something in list`.